This typo meant that 64bit floats would have their type set as a 32bit float, leading to incorrect value conversion, which then lead to Butterworth lowpass filter not getting the correct value for the frequency parameter.
git-svn-id: http://ttblue.googlecode.com/svn/branches/active@246 bc2ad3ca-9a4d-0410-8d14-35da9c29b421

+++ source/core/TTValue.cpp
@@ -32,7 +32,7 @@ TTValue::TTValue(TTFloat32 initialValue)
 TTValue::TTValue(TTFloat64 initialValue)
 {
 	data = new DataValue;
-	data->float32 = initialValue;
+	data->float64 = initialValue;
 	type = new TTDataType;
 	*type = kTypeFloat64;
 	numValues = 1;
